revert MongoService, add publishing house mongo impl
