--- app/models/Entity.scala
@@ -1,5 +1,0 @@
-package models
-
-trait Entity {
-  val collectionName: String
-}
