+++ app/models/Friend.scala
@@ -7,9 +7,7 @@ case class Friend(id: Option[Long],
                   email: Option[String],
                   comment: Option[String])
 
-object Friend extends Entity {
-
-  override val collectionName = "friends"
+object Friend {
 
   def apply(id: Long,
             fio: String,
